refactor: Shared models: Move DTO to their own project
