+++ DatabaseService/DatabaseService.csproj
@@ -34,6 +34,7 @@
 
     <ItemGroup>
       <ProjectReference Include="..\Monitoring\Monitoring.csproj" />
+      <ProjectReference Include="..\SharedModels\SharedModels.csproj" />
     </ItemGroup>
 
 </Project>
